+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 
 - [FIXED] Setting a location externally now properly cancels geolocation when still locating
 - [FIXED] Setting a location externally now properly triggers a search
+- [ADDED] Searching for locations that only have geometry data available will now also show a marker at the geometry center.
 
 ## [4.1.0] - 2019-10-10